pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "api-tester"
version = "0.1.0"
description = "Generic API testing toolkit"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
